pyproject.toml
[tool.poetry]
name = "mppiisaac"
version = "0.1.0"
description = ""
authors = [
  "Corrado Pezzato <c.pezzato@tudelft.nl>",
  "Chadi Salmi <c.salmi@tudelft.nl",
  "Elia Trevisan <e.trevisan@tudelft.nl>",
  "Max Spahn <m.spahn@tudelft.nl>",
]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8,<3.9"
pytest = "^7.2.1"
isaacgym = {path = "thirdparty/isaacgym/python", develop = true}
hydra-core = "^1.3.1"
ghalton = "^0.6.2"
mpscenes = "^0.3.1"
fabrics = "^0.6.2"
yourdfpy = "^0.0.52"
pytorch3d = "^0.3.0"
urdfenvs = "^0.7.1"
casadi = "3.5.5"
zerorpc = "^0.6.3"
matplotlib = "3.3.4"
pandas = "1.3.4"

[tool.poetry.group.dev]
optional = true

[tool.poetry.group.dev.dependencies]
pylint = "^2.16.1"
black = "^23.1.0"

[tool.poetry.group.benchmark]
optional = true

[tool.poetry.group.benchmark.dependencies]
plannerbenchmark = {version = "^1.1.1", extras = ["mpc"]}

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
